Converts `KeyTool.js` to TypesScript
- Converts `KeyTool.js` to TypeScript
- Removes JdkInstaller.js, as it's not being used and is outdated.
- Removes AndroidSdkInstaller.js, as it's not being used and is
  outdated.
- Moves code prompting for user input to outside `KeyTool.js` and
  init `init.js`.
